pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["uglyboy <uglyboy@uglyboy.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120 # YOLO
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]

ignore = [
    "E501", # line too long, handled by black
    "C901", # too complex
]